+++ src/main/java/com/azer467/saphyre/objects/items/GiftItemBase.java
@@ -1,6 +1,7 @@
 package com.azer467.saphyre.objects.items;
 
 import com.azer467.saphyre.player.gifting.GiftTables;
+import com.azer467.saphyre.player.gifting.ProbabilityList;
 import com.azer467.saphyre.util.interfaces.IGiftBase;
 import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.ItemStack;
@@ -11,18 +12,29 @@ import net.minecraft.world.World;
 import org.jetbrains.annotations.NotNull;
 
 public class GiftItemBase extends ItemBase {
-    public GiftItemBase(String name) {
+    private ProbabilityList dropList;
+
+    public GiftItemBase(String name, ProbabilityList dropList) {
         super(name);
+        setDropList(dropList);
     }
 
     @Override
     public @NotNull ActionResult<ItemStack> onItemRightClick(@NotNull World worldIn, @NotNull EntityPlayer playerIn, @NotNull EnumHand handIn) {
-        IGiftBase randomItem = GiftTables.getInstance().dropList.sample();
+        IGiftBase randomItem = dropList.sample();
         boolean handleSuccess = randomItem.executeHandler(worldIn, playerIn, handIn);
         if (handleSuccess) {
             return new ActionResult<>(EnumActionResult.PASS, playerIn.getHeldItem(handIn));
         } else {
             return new ActionResult<>(EnumActionResult.FAIL, playerIn.getHeldItem(handIn));
         }
+    }
+
+    public ProbabilityList getDropList() {
+        return dropList;
+    }
+
+    public void setDropList(ProbabilityList dropList) {
+        this.dropList = dropList;
     }
 }